backend/requirements.txt
Django>=5.2.4
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
python-docx>=1.2.0
pypdf>=5.9.0
pdf2docx>=0.5.6
yt-dlp>=2025.7.21
lxml>=6.0.0
gunicorn
python-dotenv
docx2pdf
fpdf
ffmpeg-python
Pillow>=10.0.0
rembg
onnxruntime
